src/test/java/seedu/address/logic/InternshipLogicManagerTest.java
package seedu.address.logic;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static seedu.address.logic.InternshipMessages.MESSAGE_INVALID_INTERNSHIP_DISPLAYED_INDEX;
import static seedu.address.logic.InternshipMessages.MESSAGE_UNKNOWN_COMMAND;
import static seedu.address.testutil.Assert.assertThrows;
import static seedu.address.testutil.TypicalInternships.AMY;

import java.io.IOException;
import java.nio.file.AccessDeniedException;
import java.nio.file.Path;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.io.TempDir;

import seedu.address.logic.commands.CommandResult;
import seedu.address.logic.commands.InternshipListCommand;
import seedu.address.logic.commands.exceptions.CommandException;
import seedu.address.logic.parser.exceptions.ParseException;
import seedu.address.model.InternshipModel;
import seedu.address.model.InternshipModelManager;
import seedu.address.model.InternshipUserPrefs;
import seedu.address.model.ReadOnlyInternshipData;
import seedu.address.model.internship.Internship;
import seedu.address.storage.InternshipStorageManager;
import seedu.address.storage.JsonInternshipDataStorage;
import seedu.address.storage.JsonInternshipUserPrefsStorage;
import seedu.address.testutil.InternshipBuilder;
import seedu.address.testutil.InternshipUtil;

public class InternshipLogicManagerTest {
    private static final IOException DUMMY_IO_EXCEPTION = new IOException("dummy IO exception");
    private static final IOException DUMMY_AD_EXCEPTION = new AccessDeniedException("dummy access denied exception");

    @TempDir
    public Path temporaryFolder;

    private final InternshipModel model = new InternshipModelManager();
    private InternshipLogic logic;

    @BeforeEach
    public void setUp() {
        JsonInternshipDataStorage internshipDataStorage =
                new JsonInternshipDataStorage(temporaryFolder.resolve("internshipdata.json"));
        JsonInternshipUserPrefsStorage userPrefsStorage =
                new JsonInternshipUserPrefsStorage(temporaryFolder.resolve("userPrefs.json"));
        InternshipStorageManager storage = new InternshipStorageManager(internshipDataStorage, userPrefsStorage);
        logic = new InternshipLogicManager(model, storage);
    }

    @Test
    public void execute_invalidCommandFormat_throwsParseException() {
        String invalidCommand = "uicfhmowqewca";
        assertParseException(invalidCommand, MESSAGE_UNKNOWN_COMMAND);
    }

    @Test
    public void execute_commandExecutionError_throwsCommandException() {
        String deleteCommand = "delete 9";
        assertCommandException(deleteCommand, MESSAGE_INVALID_INTERNSHIP_DISPLAYED_INDEX);
    }

    @Test
    public void execute_validCommand_success() throws Exception {
        String listCommand = InternshipListCommand.COMMAND_WORD;
        assertCommandSuccess(listCommand, InternshipListCommand.MESSAGE_SUCCESS, model);
    }

    @Test
    public void execute_storageThrowsIoException_throwsCommandException() {
        assertCommandFailureForExceptionFromStorage(DUMMY_IO_EXCEPTION, String.format(
                InternshipLogicManager.FILE_OPS_ERROR_FORMAT, DUMMY_IO_EXCEPTION.getMessage()));
    }

    @Test
    public void execute_storageThrowsAdException_throwsCommandException() {
        assertCommandFailureForExceptionFromStorage(DUMMY_AD_EXCEPTION, String.format(
                InternshipLogicManager.FILE_OPS_PERMISSION_ERROR_FORMAT, DUMMY_AD_EXCEPTION.getMessage()));
    }

    @Test
    public void getFilteredInternshipList_modifyList_throwsUnsupportedOperationException() {
        assertThrows(UnsupportedOperationException.class, () -> logic.getFilteredInternshipList().remove(0));
    }

    /**
     * Executes the command and confirms that
     * - no exceptions are thrown <br>
     * - the feedback message is equal to {@code expectedMessage} <br>
     * - the internal model manager state is the same as that in {@code expectedModel} <br>
     * @see #assertCommandFailure(String, Class, String, InternshipModel)
     */
    private void assertCommandSuccess(String inputCommand, String expectedMessage,
                                      InternshipModel expectedModel) throws CommandException, ParseException {
        CommandResult result = logic.execute(inputCommand);
        assertEquals(expectedMessage, result.getFeedbackToUser());
        assertEquals(expectedModel, model);
    }

    /**
     * Executes the command, confirms that a ParseException is thrown and that the result message is correct.
     * @see #assertCommandFailure(String, Class, String, InternshipModel)
     */
    private void assertParseException(String inputCommand, String expectedMessage) {
        assertCommandFailure(inputCommand, ParseException.class, expectedMessage);
    }

    /**
     * Executes the command, confirms that a CommandException is thrown and that the result message is correct.
     * @see #assertCommandFailure(String, Class, String, InternshipModel)
     */
    private void assertCommandException(String inputCommand, String expectedMessage) {
        assertCommandFailure(inputCommand, CommandException.class, expectedMessage);
    }

    /**
     * Executes the command, confirms that the exception is thrown and that the result message is correct.
     * @see #assertCommandFailure(String, Class, String, InternshipModel)
     */
    private void assertCommandFailure(String inputCommand, Class<? extends Throwable> expectedException,
            String expectedMessage) {
        InternshipModel expectedModel = new InternshipModelManager(model.getInternshipData(),
                new InternshipUserPrefs());
        assertCommandFailure(inputCommand, expectedException, expectedMessage, expectedModel);
    }

    /**
     * Executes the command and confirms that
     * - the {@code expectedException} is thrown <br>
     * - the resulting error message is equal to {@code expectedMessage} <br>
     * - the internal model manager state is the same as that in {@code expectedModel} <br>
     * @see #assertCommandSuccess(String, String, InternshipModel)
     */
    private void assertCommandFailure(String inputCommand, Class<? extends Throwable> expectedException,
            String expectedMessage, InternshipModel expectedModel) {
        assertThrows(expectedException, expectedMessage, () -> logic.execute(inputCommand));
        assertEquals(expectedModel, model);
    }

    /**
     * Tests the InternshipLogic component's handling of an {@code IOException} thrown by the Storage component.
     *
     * @param e the exception to be thrown by the Storage component
     * @param expectedMessage the message expected inside exception thrown by the InternshipLogic component
     */
    private void assertCommandFailureForExceptionFromStorage(IOException e, String expectedMessage) {
        Path prefPath = temporaryFolder.resolve("ExceptionInternshipUserPrefs.json");

        // Inject InternshipLogicManager with an InternshipDataStorage that throws the IOException e when saving
        JsonInternshipDataStorage internshipDataStorage = new JsonInternshipDataStorage(prefPath) {
            @Override
            public void saveInternshipData(ReadOnlyInternshipData internshipData, Path filePath)
                    throws IOException {
                throw e;
            }
        };

        JsonInternshipUserPrefsStorage userPrefsStorage =
                new JsonInternshipUserPrefsStorage(temporaryFolder.resolve("ExceptionInternshipUserPrefs.json"));
        InternshipStorageManager storage = new InternshipStorageManager(internshipDataStorage, userPrefsStorage);

        logic = new InternshipLogicManager(model, storage);

        // Triggers the saveAddressBook method by executing an add command
        String addCommand = InternshipUtil.getAddCommand(new InternshipBuilder(AMY).build());
        Internship expectedInternship = new InternshipBuilder(AMY).build();
        InternshipModelManager expectedModel = new InternshipModelManager();
        expectedModel.addInternship(expectedInternship);
        assertCommandFailure(addCommand, CommandException.class, expectedMessage, expectedModel);
    }
}
